+++ src/js/Polygon.js
@@ -18,7 +18,8 @@
  * @modified 2020-10-31 Added the `getVertexAt` function.
  * @modified 2020-11-06 Added the `move` function.
  * @modified 2020-11-10 Added the `getBounds` function.
- * @version 1.5.0
+ * @modified 2020-11-11 Generalized `move(Vertex)` to `move(XYCoords)`.
+ * @version 1.5.1
  *
  * @file Polygon
  * @public
@@ -63,6 +64,11 @@ var Polygon = /** @class */ (function () {
      * Get the polygon vertex at the given position (index).
      *
      * The index may exceed the total vertex count, and will be wrapped around then (modulo).
+     *
+     * For k >= 0:
+     *  - getVertexAt( vertices.length )     == getVertexAt( 0 )
+     *  - getVertexAt( vertices.length + k ) == getVertexAt( k )
+     *  - getVertexAt( -k )                  == getVertexAt( vertices.length -k )
      *
      * @metho getVertexAt
      * @param {number} index - The index of the desired vertex.