+++ src/js/VertTuple.d.ts
@@ -6,10 +6,12 @@
  * @author Ikaros Kappler
  * @date   2020-03-24
  * @modified 2020-05-04 Fixed a serious bug in the pointDistance function.
- * @modofied 2020-05-12 The angle(line) param was still not optional. Changed that.
- * @version 1.0.1
+ * @modified 2020-05-12 The angle(line) param was still not optional. Changed that.
+ * @modified 2020-11-11 Generalized the `add` and `sub` param from `Vertex` to `XYCoords`.
+ * @version 1.0.2
  */
 import { Vertex } from "./Vertex";
+import { XYCoords } from "./interfaces";
 export declare class VertTuple<T extends VertTuple<T>> {
     /**
      * @member {Vertex}
@@ -55,22 +57,22 @@ export declare class VertTuple<T extends VertTuple<T>> {
      * Substract the given vertex from this line's end points.
      *
      * @method sub
-     * @param {Vertex} amount The amount (x,y) to substract.
+     * @param {XYCoords} amount The amount (x,y) to substract.
      * @return {VertTuple} this
      * @instance
      * @memberof VertTuple
      **/
-    sub(amount: Vertex): VertTuple<T>;
+    sub(amount: XYCoords): VertTuple<T>;
     /**
      * Add the given vertex to this line's end points.
      *
      * @method add
-     * @param {Vertex} amount The amount (x,y) to add.
+     * @param {XYCoords} amount The amount (x,y) to add.
      * @return {Line} this
      * @instance
      * @memberof VertTuple
      **/
-    add(amount: Vertex): VertTuple<T>;
+    add(amount: XYCoords): VertTuple<T>;
     /**
      * Normalize this line (set to length 1).
      *