+++ src/js/VertTuple.js
@@ -7,8 +7,9 @@
  * @author Ikaros Kappler
  * @date   2020-03-24
  * @modified 2020-05-04 Fixed a serious bug in the pointDistance function.
- * @modofied 2020-05-12 The angle(line) param was still not optional. Changed that.
- * @version 1.0.1
+ * @modified 2020-05-12 The angle(line) param was still not optional. Changed that.
+ * @modified 2020-11-11 Generalized the `add` and `sub` param from `Vertex` to `XYCoords`.
+ * @version 1.0.2
  */
 Object.defineProperty(exports, "__esModule", { value: true });
 var Vertex_1 = require("./Vertex");
@@ -54,7 +55,7 @@ var VertTuple = /** @class */ (function () {
      * Substract the given vertex from this line's end points.
      *
      * @method sub
-     * @param {Vertex} amount The amount (x,y) to substract.
+     * @param {XYCoords} amount The amount (x,y) to substract.
      * @return {VertTuple} this
      * @instance
      * @memberof VertTuple
@@ -69,7 +70,7 @@ var VertTuple = /** @class */ (function () {
      * Add the given vertex to this line's end points.
      *
      * @method add
-     * @param {Vertex} amount The amount (x,y) to add.
+     * @param {XYCoords} amount The amount (x,y) to add.
      * @return {Line} this
      * @instance
      * @memberof VertTuple